Massive Changes

+++ webpack-plugin/electron-live-reload.ts
@@ -3,65 +3,78 @@ import { spawn, SpawnOptions, ChildProcess } from 'child_process';
 import { resolve } from 'path';
 import { Compiler, Stats } from 'webpack';
 
-interface ElectronLiveReloadPluginOptions
-{
+interface ElectronLiveReloadPluginOptions {
   test: RegExp;
   path: string;
   args?: string[];
-  options?: SpawnOptions
+  options?: SpawnOptions;
 }
 
-export default class ElectronLiveReloadPlugin
-{
+export default class ElectronLiveReloadPlugin {
   options: ElectronLiveReloadPluginOptions;
   child: ChildProcess;
   hashIndex: number[] = [];
+  persistence: (number, string) => void;
+  first: Boolean = true;
 
-  constructor(options: ElectronLiveReloadPluginOptions)
-  {
+  constructor(options: ElectronLiveReloadPluginOptions) {
     this.options = options;
     this.options.args = this.options.args || [];
     this.options.options = this.options.options || {};
     this.options.options.stdio = this.options.options.stdio || 'inherit';
+    this.persistence = this.persistencePreBind.bind(this);
   }
 
-  launch() {
-    if (this.child)
-    {
-      this.child.kill();
-      this.child = undefined;
+  persistencePreBind(code, signal) {
+    console.log('Electron process exited with code: ' + code);
+
+    if (signal) {
+      console.log('Signal given was: ' + signal);
     }
 
     this.child = spawn(
       electron as any,
       this.options.args.concat(this.options.path),
       this.options.options
-    )
+    );
+
+    this.child.on('exit', this.persistence);
   }
 
   apply(compiler: Compiler) {
-    compiler.plugin("done", (stats) => {
-      let relaunch = false;
-      stats.compilation.modules.forEach(module => {
-        if (!module.resource) return true;
-        if (!module._cachedSource) return true;
 
-        const hash = module._cachedSource.hash;
-        const split = module.portableId.split('!');
-        const id = split[split.length - 1];
+    compiler.plugin('after-emit', (compilation, callback) => {
+      const changes = compilation.modules.filter(module => {
+        if (module.resource && module._cachedSource) {
+          const hash = module._cachedSource.hash;
+          const id = module.id.split('!').pop();
+          const old = this.hashIndex[module.resource];
+          this.hashIndex[module.resource] = hash;
+          return (hash !== old) && id.match(this.options.test);
+        } else {
+          return false;
+        }
+      }).length > 0;
 
-        if (id.match(this.options.test)
-        && this.hashIndex[module.resource] !== hash)
-        {
-          relaunch = true;
+      if (changes || this.first) {
+        this.first = false;
+
+        if (this.child) {
+          this.child.removeListener('exit', this.persistence);
+          this.child.kill();
         }
 
-        this.hashIndex[module.resource] = hash;
-      })
+        console.log('Starting Electron application');
+        this.child = spawn(
+          electron as any,
+          this.options.args.concat(this.options.path),
+          this.options.options
+        );
 
-      if (relaunch) {
-        this.launch();
+        this.child.on('exit', this.persistence);
       }
-    })
+
+      callback();
+    });
   }
 }
